+++ docker-run.sh
@@ -9,14 +9,8 @@ docker run -d --name="Zoneminder" \
 -e TZ="America/New_York" \
 -e PUID="99" \
 -e PGID="100" \
--e INSTALL_HOOK="0" \
--e INSTALL_FACE="0" \
--e INSTALL_TINY_YOLOV3="0" \
--e INSTALL_YOLOV3="0" \
--e INSTALL_TINY_YOLOV4="0" \
--e INSTALL_YOLOV4="0" \
 -e MULTI_PORT_START="0" \
 -e MULTI_PORT_END="0" \
 -v "/mnt/cache/appdata/Zoneminder":"/config":rw \
 -v "/mnt/cache/appdata/Zoneminder/data":"/var/cache/zoneminder":rw \
-dlandon/zoneminder
+zoneminder