+++ zmeventnotification/zmes_hook_helpers/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "6.1.7"
+VERSION=__version__