+++ zmeventnotification/zmeventnotification/zmes_hook_helpers/__init__.py
@@ -1,2 +1,0 @@
-__version__ = "6.1.5"
-VERSION=__version__